refactor: 입력값 검증 컨트롤러에서 하도록 변경
